Move files - CustomButton, LinkCard and MessageCard
